python/pyproject.toml
[build-system]
requires = ["maturin>=1.5,<2"]
build-backend = "maturin"

[project]
name = "phantomat"
version = "0.0.1"
requires-python = ">=3.10"
dependencies = ["numpy>=1.24"]

[project.optional-dependencies]
dev = ["pytest>=8", "pytest-benchmark", "hypothesis", "Pillow"]

[tool.maturin]
manifest-path = "../crates/phantomat-python/Cargo.toml"
module-name = "phantomat._native"
python-source = "."
exclude = [
  "/__pycache__",
  "/.pytest_cache",
  "*.pyc",
  "/.venv",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
